% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{athena_supported_types}
\alias{athena_supported_types}
\title{Retrieve the current supported type names in Athena}
\usage{
athena_supported_types()
}
\description{
Retrieve the current supported type names in Athena
}
\examples{
athena_supported_types()
}
