% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis-jars-package.R
\docType{package}
\encoding{UTF-8}
\name{metis.jars}
\alias{metis.jars}
\title{Amazon Athena JDBC Driver Wrapper Supporting the 'metis' Package}
\description{
Contents of the Amazon Athena JDBC Driver
\url{https://docs.aws.amazon.com/athena/latest/ug/connect-with-jdbc.html}
provided to support functions in the 'metis' package. Version number
reflects the version number of the included 'JAR' file.
}
\section{IMPORTANT}{


Until a release of 'rJava' newer than '0.9.10' is on CRAN signals sent
from interrupting Athena JDBC calls crash the R interpreter. You therefore
need to set the \code{-Xrs} option to avoid signals being passed on to the JVM
owner. That has to be done \emph{before} \code{rJava} is loaded so you either need
to remember to put it at the top of all scripts \emph{or} stick this
in your local  \verb{~/.Rprofile} and/or sitewide \code{Rprofile}:\preformatted{options(
  "java.parameters" = c(getOption("java.parameters", default = NULL), "-Xrs")
)
}
}

\references{
\href{https://s3.amazonaws.com/athena-downloads/drivers/JDBC/SimbaAthenaJDBC_2.0.7/docs/Simba+Athena+JDBC+Driver+Install+and+Configuration+Guide.pdf}{Simba Athena JDBC Driver with SQL Connector Installation and Configuration Guide}
}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
