% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis_jar_path.R
\name{metis_jar_path}
\alias{metis_jar_path}
\title{Retrieve the fully-qualified path to the included Athena JAR file}
\usage{
metis_jar_path()
}
\description{
Retrieve the fully-qualified path to the included Athena JAR file
}
\examples{
metis_jar_path()
}
