% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{simba_driver_version}
\alias{simba_driver_version}
\title{Retrieve the version of the Simba Athena JDBC driver included with the package}
\usage{
simba_driver_version()
}
\description{
Retrieve the version of the Simba Athena JDBC driver included with the package
}
\examples{
simba_driver_version()
}
